fix: adds attribute value is empty check
