+++ dhis-2/dhis-services/dhis-service-core/src/main/java/org/hisp/dhis/deduplication/DefaultDeduplicationService.java
@@ -149,7 +149,8 @@ public class DefaultDeduplicationService
         Set<TrackedEntityAttributeValue> trackedEntityAttributeValueB = trackedEntityInstanceB
             .getTrackedEntityAttributeValues();
 
-        if ( null == trackedEntityAttributeValueA || null == trackedEntityAttributeValueB )
+        if ( ( null == trackedEntityAttributeValueA || trackedEntityAttributeValueA.isEmpty() ) ||
+                ( null == trackedEntityAttributeValueB || trackedEntityAttributeValueB.isEmpty() ))
         {
             return true;
         }
